# faster min/max for IEEEFloats