Faster `min`, `max`, and `minmax` for `IEEEFloat` (`Float16`, `Float32`, `Float64`) types.

Setup:
```julia
using BenchmarkTools

a = randn(1000);
b = randn(size(a));
z = similar(a);
zz = Array{NTuple{2,Float64}}(undef,size(a));

@btime map!(min, $z, $a, $b);
@btime map!(max, $z, $a, $b);
@btime map!(minmax, $zz, $a, $b);

@btime broadcast!(min, $z, $a, $b);
@btime broadcast!(max, $z, $a, $b);
@btime broadcast!(minmax, $zz, $a, $b);
```

Before:
```
  1.790 μs (0 allocations: 0 bytes)
  1.780 μs (0 allocations: 0 bytes)
  2.244 μs (0 allocations: 0 bytes)
  308.502 ns (0 allocations: 0 bytes)
  307.631 ns (0 allocations: 0 bytes)
  1.970 μs (0 allocations: 0 bytes)
```

After:
```
  1.240 μs (0 allocations: 0 bytes)
  1.240 μs (0 allocations: 0 bytes)
  1.790 μs (0 allocations: 0 bytes)
  227.292 ns (0 allocations: 0 bytes)
  227.292 ns (0 allocations: 0 bytes)
  552.128 ns (0 allocations: 0 bytes)
```

`map!` indicates the straight-line performance while `broadcast!` allows for SIMD. This PR increases throughput for `min`/`max` by about 35%, `minmax` by 25%, and enables `minmax` to SIMD (additional 225% throughput).

Note that the `AbstractFloat`-specialized `minmax` was deleted and the `Real` version redefined to simply `minmax(x,y) = min(x,y),max(x,y)`. The new version is faster for `IEEEFloat` (with the `min`/`max` changes), does not change performance for `BitInteger`, and does not apply to `BigFloat`.

The `AbstractFloat`-specialized versions of `min`/`max` are now dead to `Base` types (`BigFloat` has its own specialization and `IEEEFloat` is superseded by the new specializations in this PR). I was tempted to remove these methods, but that would be breaking for any user-defined `AbstractFloat` types that rely on them. I'm not sure that any do, but I didn't feel the need to break things in this PR.