@oscardssmith 2022-06-01
Can you give me a test of this on every pair of `Float16`? Other than that (just to be sure this handles all the edge cases), this looks great.