@N5N3 2022-06-01
My locate bench shows: (Compared with #41709)
`Float32/Float64`: this PR has faster `map!`, but slight slower `broadcast!`.
`Float16`: this PR seems slower on both case. (My PC has no Native `Float16` support)