@mcabbott 2022-06-01
Timing this and https://github.com/JuliaLang/julia/pull/41709 on M1 mac.
* This has no effect on `map!` speed, while 41709 takes 60% as long
* Both improve `broadcast!`, this one better: 73% as long, vs  41709 at 79% as long.
Same ratios for Float64 and Float32.