@N5N3 2022-06-01
@mcabbott would you mind bench this version?
```julia
function __min(x, y)
    diff = x - y
    min = ifelse(signbit(diff), x, y)
    anynan = isnan(x)|isnan(y)
    ifelse(anynan, diff, min)
end

function __max(x, y)
    diff = x - y
    max = ifelse(signbit(diff), y, x)
    anynan = isnan(x)|isnan(y)
    ifelse(anynan, diff, max)
end

function __minmax(x, y)
    diff = x - y
    sdiff = signbit(diff)
    min = ifelse(sdiff, x, y)
    max = ifelse(sdiff, y, x)
    anynan = isnan(x)|isnan(y)
    min = ifelse(anynan, diff, min)
    max = ifelse(anynan, diff, max)
    min, max
end
```