@mcabbott 2022-06-01
On M1 `__min` & `__max` take
* the same 60% time for `map!` as 41709
* 62% for Float64 and 52% for Float32 in  `broadcast!`.