@mikmoore 2022-06-01
Testing on every `Float16`:
```julia
function makequiet(x)
	# set the quiet bit of a NaN
	u = reinterpret(Unsigned, x)
	qbit = one(u) << (Base.significand_bits(typeof(x))-1) # quiet bit in NaN
	return reinterpret(typeof(x), u | qbit)
end

i16s = typemin(Int16):typemax(Int16)
for ix in i16s, iy in i16s
	x = reinterpret(Float16, ix)
	y = reinterpret(Float16, iy)
	m = min(x, y)
	if x < y
		m == x || println(x,",",y," -> ",m)
	elseif x > y
		m == y || println(x,",",y," -> ",m)
	elseif x == y
		m == x == y && signbit(m) == signbit(x) | signbit(y) || println(x,",",y," -> ",m)
	elseif isnan(x) | isnan(y)
		makequiet(m) === makequiet(x) || makequiet(m) === makequiet(y) || println(x,",",y," -> ",m," : ",ix,",",iy," -> ",reinterpret(Int16,m))
	end
end
```
which took 25 minutes on my machine. I did a similar test for `max`. Neither produced any notices.

My machine lacks native `Float16` support so `Float16` is tragically slow. Without native support, it's likely that many `Float16` operations will be slow and not just the ones here. So I don't know how critical that performance is.

I suppose the faster approach would be to promote to `Float32` for non-native `Float16`. However, it would seem better to make the fundamental operations (`<`, `>`, `+`, `isnan`) faster for `Float16` (if only by promotion) and then the performance here would be less horrid.