@mikmoore 2022-06-01
Non-native `Float16` performance should not be a priority, in my opinion. Someone without native `Float16` is unlikely to use them and will suffer performance everywhere if they do. I imagine there's no qualitative difference when using native `Float16` as compared to native `Float32`/`Float64`, since all candidates here are based on primitive float and logical instructions.

Looking at `Float64`, I see better performance of this PR's `min` on `map!` and better for `__min` on `broadcast!`. The discrepancy is easy looking at the `code_native`. On my Intel machine, this `min` is branchless in both cases, whereas `__min` uses one branch in the `map` (and scalar) case but manages to replace this with a `vblendvpd` when `broadcast` (SIMD).

I'm torn. Scalar operation is probably more common in the places where performance is critical, but ultimately the difference won't add up to anything massive. Vectorized operation gives the potential to magnify a small performance difference, but again the `min`/`max` call is likely only a small part of anything but a micro-benchmark. In either case, the difference is only 10-20%.

I tried to tweak `__min` to convince it to drop the branch, but couldn't. If someone can manage to do that, that would likely yield the best implementation.