@mikmoore 2022-06-01
Certainly. I was referring to the other line of discussion looking at `Float16` performance.