@N5N3 2022-06-01
> I have a different approach (related to the isless implementation) for float minimum/maximum that is more than 4x faster than what exists now and should avoid the integer slowdown. I will finalize it and open a PR when I get a chance.

That's great! I tried more thorough unroll to make `minimum` faster in #43725. But it increases TTFP. Hope we can close https://github.com/JuliaLang/julia/issues/31442 this time.