@N5N3 2022-06-02
LLVM already has a issue to track X86's `minimum`/`maximum`. https://github.com/llvm/llvm-project/issues/53353
Since the optimizaiton is somehow platform dependent, I believe the best choice is use LLVM's `{min,max}imum` when it's available.
This should also enable auto-vectorlization and make `foldl(min, a)` as fast as `reduce(min, a)`.