@mikmoore 2022-06-15
Having spent some time thinking about it, I think that the `__min` and `__max` proposed by @N5N3 are the best candidates. Those versions performed better than the original versions proposed in this PR in some situations and almost as well in others. But I think that if LLVM eventually manages to produce the code I wish it would [https://discourse.julialang.org/t/i-think-llvm-can-do-better-here/82213](url), these will be indisputably the best.

It'd be nice if we can just use `minmax(x,y) = min(x,y),max(x,y)` and rely on inlining to deliver full efficiency, but we'll need to see whether that happens or if a manual definition is required.

@N5N3, you are managing #41709 which accomplishes everything this PR would. Would you like me to close this and you can move forward with that one or should I adopt your versions and continue this PR?